+++ frontend/src/router.ts
@@ -2,8 +2,48 @@ import Vue from 'vue';
 import Router from 'vue-router';
 
 import RouterComponent from './components/RouterComponent.vue';
+import {useUserStore} from "@/stores/user";
+import {useMainStore} from "@/stores/main";
+
+
+async function routeGuard(to, from, next) {
+    const userStore = useUserStore();
+    const mainStore = useMainStore();
+
+    if (!mainStore.license) {
+        await mainStore.fetchLicense();
+    }
+    if (!mainStore.license?.active) {
+        if (to.path !== '/setup') {
+            next('/setup');
+        } else {
+            next();
+        }
+    } else {
+        await userStore.checkLoggedIn();
+
+        if (userStore.isLoggedIn && (to.path === '/auth/login' || to.path === '/')) {
+            next('/main/dashboard');
+        } else if (!userStore.isLoggedIn && (to.path === '/' || (to.path as string).startsWith('/main') || (to.path as string).startsWith('/setup'))) {
+            next('/auth/login');
+        } else if (to.path.startsWith('/main/admin') && !userStore.hasAdminAccess) {
+            next('/main');
+        } else {
+            next();
+        }
+    }
+}
+
 
 Vue.use(Router);
+Vue.mixin({
+    async beforeRouteEnter(to, from, next) {
+        await routeGuard(to, from, next);
+    },
+    async beforeRouteUpdate(to, from, next) {
+        await routeGuard(to, from, next);
+    }
+})
 
 export default new Router({
     mode: 'history',
@@ -182,6 +222,7 @@ export default new Router({
                             ],
                         },
                         {
+                            name: 'admin',
                             path: 'admin',
                             component: () => import(/* webpackChunkName: "main-admin" */ './views/main/admin/Admin.vue'),
                             redirect: 'admin/general',
